# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: AI-powered expense tracking system with SMS webhook integration
  version: 1.0.0
  contact:
    name: API Support
    
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:3049
    description: Production server (Docker)

tags:
  - name: Categories
    description: Manage income and expense categories
  - name: Transactions
    description: Manage financial transactions
  - name: Webhooks
    description: SMS webhook endpoints
  - name: Health
    description: Service health checks

paths:
  /health:
    get:
      summary: Health check
      tags: [Health]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Expense Tracker API is running

  /api/ping:
    get:
      summary: Ping endpoint
      tags: [Health]
      responses:
        '200':
          description: Pong response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: pong

  /api/categories:
    get:
      summary: List all categories
      tags: [Categories]
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new category
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  example: Groceries
                type:
                  type: string
                  enum: [income, expense]
                  example: expense
                icon:
                  type: string
                  example: 🛒
                color:
                  type: string
                  example: "#4CAF50"
                monthly_budget:
                  type: number
                  format: double
                  example: 5000.00
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Category already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/categories/{id}:
    get:
      summary: Get category by ID
      tags: [Categories]
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/transactions:
    get:
      summary: List transactions
      tags: [Transactions]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: source
          in: query
          schema:
            type: string
            enum: [sms, email, bank_statement, manual]
      responses:
        '200':
          description: Paginated list of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  page:
                    type: integer
                  limit:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create manual transaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - type
                - description
                - transaction_date
              properties:
                amount:
                  type: number
                  format: double
                  minimum: 0.01
                  example: 420.00
                type:
                  type: string
                  enum: [income, expense]
                  example: expense
                description:
                  type: string
                  example: Lunch at restaurant
                transaction_date:
                  type: string
                  format: date-time
                  example: "2026-01-08T12:30:00Z"
                category_id:
                  type: string
                  format: uuid
                  nullable: true
                merchant:
                  type: string
                  nullable: true
                  example: Swiggy
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/transactions/review:
    get:
      summary: Get transactions requiring review
      tags: [Transactions]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Transactions needing review
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  count:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/transactions/{id}:
    get:
      summary: Get transaction by ID
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update transaction
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: double
                type:
                  type: string
                  enum: [income, expense]
                description:
                  type: string
                transaction_date:
                  type: string
                  format: date-time
                category_id:
                  type: string
                  format: uuid
                  nullable: true
                merchant:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete transaction
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Transaction deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/transactions/{id}/approve:
    put:
      summary: Approve AI-extracted transaction
      description: Removes transaction from review queue and applies AI suggested category
      tags: [Transactions]
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/webhooks/sms:
    post:
      summary: Receive SMS webhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSWebhook'
      responses:
        '200':
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string
                    example: Webhook received successfully
                  id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Category UUID

    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Transaction UUID

  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Groceries
        type:
          type: string
          enum: [income, expense]
          example: expense
        icon:
          type: string
          example: 🛒
        color:
          type: string
          example: "#4CAF50"
        monthlyBudget:
          type: number
          format: double
          nullable: true
          example: 5000.00
        isSystem:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          example: Payment to APNA DHABA
        amount:
          type: number
          format: double
          example: 420.00
        type:
          type: string
          enum: [income, expense]
          example: expense
        transactionDate:
          type: string
          format: date-time
        categoryId:
          type: string
          format: uuid
          nullable: true
        category:
          $ref: '#/components/schemas/Category'
        merchant:
          type: string
          nullable: true
          example: APNA DHABA
        tags:
          type: array
          items:
            type: string
        source:
          type: string
          enum: [sms, email, bank_statement, manual]
          example: sms
        sourceId:
          type: string
          format: uuid
          nullable: true
        aiConfidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true
          example: 0.95
        aiSuggestedCategoryId:
          type: string
          format: uuid
          nullable: true
        aiSuggestedCategory:
          $ref: '#/components/schemas/Category'
        aiMetadata:
          type: string
          nullable: true
        requiresReview:
          type: boolean
          example: false
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SMSWebhook:
      type: object
      required:
        - deviceId
        - event
        - id
        - payload
        - webhookId
      properties:
        deviceId:
          type: string
          example: "_HHGhfi5jlfmB78Us7l7g"
        event:
          type: string
          example: "sms:received"
        id:
          type: string
          example: "TEST-WEBHOOK-001"
        payload:
          type: object
          properties:
            message:
              type: string
            receivedAt:
              type: string
              format: date-time
            messageId:
              type: string
            phoneNumber:
              type: string
            simNumber:
              type: integer
        webhookId:
          type: string
          example: "pixel9"

    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Transaction not found"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Failed to fetch transactions"